Combine effects test files
